finish section 73

+++ graphql-prisma/src/Resolvers/Mutation.js
@@ -96,7 +96,18 @@ const Mutation = {
             }
         }, info)
     },
-    updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) throw new Error('Unable to find post.')
+
         return prisma.mutation.updatePost({
             where: {
                 id: args.id
@@ -104,13 +115,15 @@ const Mutation = {
             data: args.data
         }, info)
     },
-    createComment(parent, args, { prisma }, info) {
+    createComment(parent, args, { prisma, request  }, info) {
+        const userId = getUserId(request)
+        
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -121,14 +134,35 @@ const Mutation = {
             }
         }, info)
     },
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) throw new Error('Unable to find comment.')
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) throw new Error('Unable to find comment.')
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
